+++ module_programming_themisml/module_programming_themisml/extract_methods/languages/java/java_method_parser_listener.py
@@ -1,7 +1,7 @@
 # Generated from JavaParser.g4 by ANTLR 4.11.1
 from antlr4 import *
 
-from .method_node import MethodNode
+from ...method_node import MethodNode
 
 JAVA_METHOD_DECLARATIONS = ["methodDeclaration", "constructorDeclaration", "genericMethodDeclaration",
                             "genericConstructorDeclaration"]
